requirements.txt
blinker==1.9.0
boto3==1.38.41
botocore==1.38.41
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.1.8
Flask==3.1.1
flask-cors==6.0.1
Flask-SQLAlchemy==3.1.1
greenlet==3.2.3
idna==3.10
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
MarkupSafe==3.0.2
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
requests==2.32.4
s3transfer==0.13.0
six==1.17.0
SQLAlchemy==2.0.41
typing_extensions==4.14.0
urllib3==1.26.20
Werkzeug==3.1.3
zipp==3.23.0
